requirements.txt
numpy
openpyxl
pandas
pdfrw
Pillow
reportlab
PyMuPdf
